add nav types
