+++ src/lib/types/index.ts
@@ -1,0 +1,2 @@
+export * from './podcast';
+export * from './routes';
