+++ src/lib/types/routes.tsx
@@ -1,0 +1,6 @@
+import {Episode} from './podcast';
+
+export type RootStackParamList = {
+  Home: undefined;
+  Episode: {episode: Episode};
+};
